Modularized the functions in the signin and Book clubhouse page.
Also created a date picker widget.
