+++ lib/pages/sign_in.dart
@@ -1,14 +1,14 @@
 // ignore_for_file: use_build_context_synchronously
 
-import 'package:booking_app/constants.dart';
+import 'package:booking_app/functions/sign_functions.dart';
 import 'package:booking_app/pages/homepage.dart';
 import 'package:booking_app/widgets/textboxes/password_box.dart';
 import 'package:booking_app/widgets/buttons/primary_button.dart';
-import 'package:booking_app/widgets/textboxes/text_box_wcontroller.dart';
 import 'package:booking_app/widgets/textboxes/text_box_wcontroller_numeric.dart';
 import 'package:booking_app/widgets/textbuttons/primary_text_button.dart';
 import 'package:flutter/material.dart';
-import 'package:cloud_firestore/cloud_firestore.dart';
+
+GlobalKey<ScaffoldState> scaffoldKey = GlobalKey<ScaffoldState>();
 
 class SignInPage extends StatefulWidget {
   @override
@@ -19,10 +19,13 @@ class _SignInPageState extends State<SignInPage> {
   final TextEditingController _villaNumberController = TextEditingController();
   final TextEditingController _passwordController = TextEditingController();
   bool _isLoading = false;
+  dynamic signInResult;
+  dynamic signFunctions = SignFunctions();
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
+      key: scaffoldKey,
       appBar: AppBar(
         automaticallyImplyLeading: false,
         backgroundColor: Colors.black,
@@ -51,7 +54,45 @@ class _SignInPageState extends State<SignInPage> {
               FractionallySizedBox(
                 widthFactor: 1,
                 child: PrimaryButton(
-                  onPressed: () => signIn(context),
+                  onPressed: () async {
+                    setState(() {
+                      _isLoading = true;
+                    });
+                    signInResult = await signFunctions.signIn(
+                        _villaNumberController,
+                        _passwordController);
+                    if (signInResult[0] == 0) {
+                      Navigator.push(
+                          context,
+                          MaterialPageRoute(
+                              builder: (context) => MyHomePage(
+                                    villa_num: int.tryParse(
+                                            _villaNumberController.text
+                                                .trim()) ??
+                                        0,
+                                    userData: signInResult[1],
+                                  )));
+                    } else if (signInResult[0] == 1) {
+                      showDialog(
+                        context: context,
+                        builder: (BuildContext context) {
+                          return AlertDialog(
+                            title: Text('Invalid Credentials'),
+                            content: Text('The villa number or password is incorrect.'),
+                            actions: [
+                              PrimaryTextButton(
+                                text: 'OK',
+                                onPressed: () => Navigator.of(context).pop(),
+                              ),
+                            ],
+                          );
+                        },
+                      );
+                    }
+                    setState(() {
+                      _isLoading = false;
+                    });
+                  },
                   text: 'Sign In',
                   isLoading: _isLoading,
                 ),
@@ -61,57 +102,5 @@ class _SignInPageState extends State<SignInPage> {
         ),
       ),
     );
-  }
-
-  Future<void> signIn(BuildContext context) async {
-    int villaNumber = int.tryParse(_villaNumberController.text.trim()) ?? 0;
-    String password = _passwordController.text.trim();
-
-    setState(() {
-      _isLoading = true;
-    });
-
-    try {
-      // Query Firestore to check if the villa number and password combination exists in "villa_users"
-      QuerySnapshot<Map<String, dynamic>> snapshot = await FirebaseFirestore
-          .instance
-          .collection(firestoreVillaUsersCollection)
-          .where('Villa_num', isEqualTo: villaNumber)
-          .where('Password', isEqualTo: password)
-          .get();
-
-      if (snapshot.docs.isNotEmpty) {
-        List<dynamic> userMaps = snapshot.docs.first.data()['userMaps'];
-        Navigator.push(
-            context,
-            MaterialPageRoute(
-                builder: (context) => MyHomePage(
-                      villa_num: villaNumber,
-                      userData: userMaps,
-                    )));
-      } else {
-        showDialog(
-          context: context,
-          builder: (BuildContext context) {
-            return AlertDialog(
-              title: Text('Invalid Credentials'),
-              content: Text('The villa number or password is incorrect.'),
-              actions: [
-                PrimaryTextButton(
-                  text: 'OK',
-                  onPressed: () => Navigator.of(context).pop(),
-                ),
-              ],
-            );
-          },
-        );
-      }
-    } catch (e) {
-      debugPrint('Error during sign-in: $e');
-    } finally {
-      setState(() {
-        _isLoading = false;
-      });
-    }
   }
 }
