+++ lib/pages/view_bookings.dart
@@ -2,13 +2,14 @@ import 'package:booking_app/constants.dart';
 import 'package:booking_app/pages/book_clubhouse.dart';
 import 'package:booking_app/widgets/buttons/view_bookings_date_button.dart';
 import 'package:booking_app/widgets/buttons/view_bookings_option_button.dart';
+import 'package:booking_app/widgets/datepicker/date_picker.dart';
 import 'package:booking_app/widgets/textbuttons/accept_text_button.dart';
-import 'package:booking_app/widgets/buttons/delete_button.dart';
 import 'package:booking_app/widgets/textbuttons/reject_text_button.dart';
-import 'package:booking_app/widgets/buttons/secondary_button.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
+
+final FirebaseFirestore firestore = FirebaseFirestore.instance;
 
 class ViewClubhouseBookings extends StatefulWidget {
   final int villaNum;
@@ -22,45 +23,14 @@ class ViewClubhouseBookings extends StatefulWidget {
 
 class _ViewClubhouseBookingsState extends State<ViewClubhouseBookings> {
   DateTime selectedDate = DateTime.now();
+  dynamic asyncDate;
   final double tablePadding = 7;
+  final customDatePicker = CustomDatePicker();
 
   @override
   void initState() {
     super.initState();
     selectedDate = widget.selected_date;
-  }
-
-  void _selectDate(BuildContext context) async {
-    final DateTime? pickedDate = await showDatePicker(
-      context: context,
-      initialDate: DateTime.now(),
-      firstDate: DateTime(DateTime.now().year - 1),
-      lastDate: DateTime(DateTime.now().year + 1),
-      builder: (BuildContext context, Widget? child) {
-        // Theme customization for the date picker
-        final ThemeData theme = Theme.of(context);
-        return Theme(
-          data: theme.copyWith(
-            colorScheme: theme.colorScheme.copyWith(
-              primary: Colors.black,
-              onPrimary: Colors.white,
-            ),
-            textTheme: theme.textTheme.copyWith(
-              titleMedium: const TextStyle(
-                color: Colors.black,
-                fontWeight: FontWeight.bold,
-              ),
-            ),
-          ),
-          child: child!,
-        );
-      },
-    );
-    if (pickedDate != null) {
-      setState(() {
-        selectedDate = pickedDate;
-      });
-    }
   }
 
   @override
@@ -75,7 +45,12 @@ class _ViewClubhouseBookingsState extends State<ViewClubhouseBookings> {
             widthFactor: 1,
             child: ViewBookingsDateButton(
                 text: DateFormat('d MMMM yyyy').format(selectedDate),
-                onPressed: () => _selectDate(context)),
+                onPressed: () async {
+                  asyncDate = await customDatePicker.selectDate(context);
+                  setState(() {
+                    selectedDate = asyncDate == null ? selectedDate : asyncDate;
+                  });
+                }),
           ),
           const SizedBox(
             height: 20,
@@ -274,46 +249,6 @@ class ClubhouseBookingDetails extends StatefulWidget {
 class _ClubhouseBookingDetails extends State<ClubhouseBookingDetails> {
   bool isLoading = false;
 
-  Future<void> deleteBooking(DocumentReference bookingRef) async {
-    setState(() {
-      isLoading = true;
-    });
-    showDialog(
-      context: context,
-      builder: (BuildContext context) {
-        return AlertDialog(
-          title: const Text('Confirm delete'),
-          content: const Text('Do you want to delete the booking?'),
-          actions: [
-            AcceptTextButton(
-              text: 'Yes',
-              onPressed: () async {
-                try {
-                  await bookingRef.delete();
-                  debugPrint('Booking deleted successfully.');
-                  Navigator.pop(context);
-                  Navigator.pop(context);
-                } catch (error) {
-                  debugPrint('Error deleting booking: $error');
-                } finally {
-                  setState(() {
-                    isLoading = false;
-                  });
-                }
-              },
-            ),
-            RejectTextButton(
-              text: 'No',
-              onPressed: () {
-                Navigator.of(context).pop();
-              },
-            ),
-          ],
-        );
-      },
-    );
-  }
-
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -325,7 +260,44 @@ class _ClubhouseBookingDetails extends State<ClubhouseBookingDetails> {
             IconButton(
                 onPressed: () {
                   if (widget.villa_number == widget.data['villa_no']) {
-                    deleteBooking(widget.bookingRef);
+                    setState(() {
+                      isLoading = true;
+                    });
+                    showDialog(
+                      context: context,
+                      builder: (BuildContext context) {
+                        return AlertDialog(
+                          title: const Text('Confirm delete'),
+                          content:
+                              const Text('Do you want to delete the booking?'),
+                          actions: [
+                            AcceptTextButton(
+                              text: 'Yes',
+                              onPressed: () async {
+                                try {
+                                  await widget.bookingRef.delete();
+                                  debugPrint('Booking deleted successfully.');
+                                  Navigator.pop(context);
+                                  Navigator.pop(context);
+                                } catch (error) {
+                                  debugPrint('Error deleting booking: $error');
+                                } finally {
+                                  setState(() {
+                                    isLoading = false;
+                                  });
+                                }
+                              },
+                            ),
+                            RejectTextButton(
+                              text: 'No',
+                              onPressed: () {
+                                Navigator.of(context).pop();
+                              },
+                            ),
+                          ],
+                        );
+                      },
+                    );
                   }
                   ;
                 },
