+++ lib/widgets/datepicker/date_picker.dart
@@ -1,0 +1,60 @@
+import 'package:flutter/material.dart';
+
+class CustomDatePicker {
+  Future<DateTime?> selectDate(BuildContext context) async {
+    final DateTime? pickedDate = await showDatePicker(
+      context: context,
+      initialDate: DateTime.now(),
+      firstDate: DateTime(DateTime.now().year - 1),
+      lastDate: DateTime(DateTime.now().year + 1),
+      builder: (BuildContext context, Widget? child) {
+        // Theme customization for the date picker
+        final ThemeData theme = Theme.of(context);
+        return Theme(
+          data: theme.copyWith(
+            colorScheme: theme.colorScheme.copyWith(
+              primary: Colors.black,
+              onPrimary: Colors.white,
+            ),
+            textTheme: theme.textTheme.copyWith(
+              titleMedium: const TextStyle(
+                color: Colors.black,
+                fontWeight: FontWeight.bold,
+              ),
+            ),
+          ),
+          child: child!,
+        );
+      },
+    );
+    return pickedDate;
+  }
+  Future<DateTime?> selectDateFromCurrentDate(BuildContext context) async {
+    final DateTime? pickedDate = await showDatePicker(
+      context: context,
+      initialDate: DateTime.now(),
+      firstDate: DateTime.now(),
+      lastDate: DateTime(DateTime.now().year + 1),
+      builder: (BuildContext context, Widget? child) {
+        // Theme customization for the date picker
+        final ThemeData theme = Theme.of(context);
+        return Theme(
+          data: theme.copyWith(
+            colorScheme: theme.colorScheme.copyWith(
+              primary: Colors.black,
+              onPrimary: Colors.white,
+            ),
+            textTheme: theme.textTheme.copyWith(
+              titleMedium: const TextStyle(
+                color: Colors.black,
+                fontWeight: FontWeight.bold,
+              ),
+            ),
+          ),
+          child: child!,
+        );
+      },
+    );
+    return pickedDate;
+  }
+}
